openAPI to TS complete
